[FIX]: eventManager, camera obj, color, const and copy constructor

+++ src/system/pointLightRenderSystem.cpp
@@ -21,13 +21,13 @@ void ven::PointLightRenderSystem::render(const FrameInfo &frameInfo) const
 void ven::PointLightRenderSystem::update(const FrameInfo &frameInfo, GlobalUbo &ubo)
 {
     const glm::mat4 rotateLight = rotate(glm::mat4(1.F), frameInfo.frameTime, {0.F, -1.F, 0.F});
-    int lightIndex = 0;
+    uint16_t lightIndex = 0;
 
     for (Light &light : frameInfo.lights | std::views::values) {
         assert(lightIndex < MAX_LIGHTS && "Too many lights");
         light.transform3D.translation = glm::vec3(rotateLight * glm::vec4(light.transform3D.translation, 1.F));
-        ubo.pointLights.at(static_cast<long unsigned int>(lightIndex)).position = glm::vec4(light.transform3D.translation, 1.F);
-        ubo.pointLights.at(static_cast<long unsigned int>(lightIndex)).color = light.color;
+        ubo.pointLights.at(lightIndex).position = glm::vec4(light.transform3D.translation, 1.F);
+        ubo.pointLights.at(lightIndex).color = light.color;
         lightIndex++;
     }
     ubo.numLights = lightIndex;
